Split misc dependencies into their own file.
